clusterStickiness

Tests for enrichment of doublets created from each cluster (i.e. cluster's stickiness). Only applicable with >=4 clusters. Note that when applied to an multisample object, this functions assumes that the cluster labels match across samples.

Usage

clusterStickiness(
  x,
  type = c("quasibinomial", "nbinom", "binomial", "poisson"),
  inclDiff = NULL,
  verbose = TRUE
)

Arguments

x

A table of double statistics, or a SingleCellExperiment on which scDblFinder was run using the cluster-based approach.

type

The type of test to use (quasibinomial recommended).

inclDiff

Logical; whether to include the difficulty in the model. If NULL, will be used only if there is a significant trend with the enrichment.

verbose

Logical; whether to print additional running information.

Value

A table of test results for each cluster.

Examples

sce <- mockDoubletSCE(rep(200,5))
sce <- scDblFinder(sce, clusters=TRUE, artificialDoublets=500)
#> Clustering cells...
#> 5 clusters
#> Creating ~500 artificial doublets...
#> Dimensional reduction
#> Evaluating kNN...
#> Training model...
#> iter=0, 88 cells excluded from training.
#> iter=1, 88 cells excluded from training.
#> iter=2, 88 cells excluded from training.
#> Threshold found:0.966
#> 88 (8.1%) doublets called
clusterStickiness(sce)
#>      Estimate Std. Error     t value   p.value       FDR
#> 2  0.34485810  0.2199772  1.56769927 0.1777369 0.8886845
#> 1  0.24933578  0.2153575  1.15777589 0.2992514 1.0000000
#> 5 -0.12849710  0.3478157 -0.36944017 0.7269296 1.0000000
#> 3  0.02665257  0.2565217  0.10389986 0.9212877 1.0000000
#> 4 -0.02213213  0.2376672 -0.09312237 0.9294226 1.0000000