getCellPairs

Given a vector of cluster labels, returns pairs of cross-cluster cells

Usage

getCellPairs(
  clusters,
  n = 1000,
  ls = NULL,
  q = c(0.1, 0.9),
  selMode = "proportional",
  soft.min = 5
)

Arguments

clusters

A vector of cluster labels for each cell, or a list containing metacells and graph

n

The number of cell pairs to obtain

ls

Optional library sizes

q

Library size quantiles between which to include cells (ignored if `ls` is NULL)

selMode

How to decide the number of pairs of each kind to produce. Either 'proportional' (default, proportional to the abundance of the underlying clusters), 'uniform' or 'sqrt'.

soft.min

Minimum number of pairs of a given type.

Value

A data.frame with the columns

Examples

# create random labels
x <- sample(head(LETTERS), 100, replace=TRUE)
getCellPairs(x, n=6)
#>    cell1 cell2 orig.clusters
#> 1     67    70           A+B
#> 2     52    17           A+C
#> 3     68    13           B+C
#> 4     93    12           A+D
#> 5     33    18           B+D
#> 6     20    12           C+D
#> 7     11    35           A+E
#> 8     74    77           B+E
#> 9     87    21           C+E
#> 10    82    77           D+E
#> 11    52    38           A+F
#> 12    90     1           B+F
#> 13    45    42           C+F
#> 14    72    38           D+F
#> 15    50    98           E+F