selFeatures

Selects features based on cluster-wise expression or marker detection, or a combination.

Usage

selFeatures(
  sce,
  clusters = NULL,
  nfeatures = 1000,
  propMarkers = 0,
  FDR.max = 0.05
)

Arguments

sce

A SummarizedExperiment-class, SingleCellExperiment-class with a 'counts' assay.

clusters

Optional cluster assignments. Should either be a vector of labels for each cell.

nfeatures

The number of features to select.

propMarkers

The proportion of features to select from markers (rather than on the basis of high expression). Ignored if `clusters` isn't given.

FDR.max

The maximum marker binom FDR to be included in the selection. (see findMarkers).

Value

A vector of feature (i.e. row) names.

Examples

sce <- mockDoubletSCE()
selFeatures(sce, clusters=sce$cluster, nfeatures=5)
#> [1] "gene80"  "gene149" "gene130" "gene186" "gene200"